dashboard/requirements.txt
dash==2.0.0
dash_auth==1.4.1
dash_bootstrap_components==1.0.0
pandas==1.3.4
plotly==5.4.0
PyMySQL==1.0.2
python-dotenv==0.21.0
